feat(system) add the DDL execution

+++ rmdb/src/system/sm_meta.h
@@ -57,7 +57,7 @@ struct IndexMeta {
 
 	friend std::istream &operator>>(std::istream &is, IndexMeta &index) {
 		is >> index.tab_name >> index.col_tot_len >> index.col_num;
-		for (int i = 0; i < index.col_num; ++i) {
+		for (size_t i = 0; i < index.col_num; ++i) {
 			ColMeta col;
 			is >> col;
 			index.cols.push_back(col);
